RF: move all code from datasets into __init__
