+++ nidata/atlas/mni152_template/mni152_template.rst
@@ -1,0 +1,1 @@
+TODO: MNI152_template needs an RST FILE.
