+++ nidata/core/__init__.py
@@ -1,0 +1,3 @@
+"""
+Common functions for dataset fetching and image utilities.
+"""
