+++ nidata/multimodal/hcp/hcp.rst
@@ -1,1 +1,1 @@
-dataset.py
+TODO: HCP needs a RST file.
